chore(1-3435): move milestone card to legacy milestone card
